Move files, edit readme
