+++ README.md
@@ -1,44 +1,23 @@
-# BigDataAssignment-2022
+# Big Data Házi feladat 2022
 
-Big Data analitics course assignment
+Készítők: Csilling Tamás és Knyihár Gábor
 
-## Tasks
-
-BKK3 (egyelőre magyarul, TODO: lefordítani)
-
-### Téma
+## Feladat (BKK3)
 
 Az I épületből este 10-kor, tömegközlekedéssel elérhető "kocsmák" (i.e., Google Maps találat erre a keresésre) elemzése
 
-### Adatbeszerzés
+[További részletek](doc/task.md)
 
-https://openmobilitydata.org/p/bkk/42 és pl. Google Places API
+## Megoldás
 
-### Adatelőkészítés és szűkítés
+### Előre lefuttatott notebook
 
-A Google Places API (a free változat elég kell, hogy legyen) talán a fő kihívás, egyenértékes megoldást elfogadok
+[Notebook](final.ipynb)
 
-### Leskálázás méretben "small" datára
+### Eredmények újra futtatása
 
-Valószínűleg nem szükséges
+```
+pip install -r requirements.txt
+jupyter notebook final.ipynb
+```
 
-### EDA fókusz
-
-Az I épület 10 perces sétatávolságában megállóval rendelkező éjszakai járatok alapvető jellemzői (javasolt:
-pandas_profiling vagy BambooLib)
-
-### Big Data vizualizáció
-
-Datashader alapú (lehet HoloViz-be integrálva) interaktív, Budapest-térkép alapú heatmap, mely az I épületből este 10-es
-indulással, átszállás nélkül való utazási időben vett távolságát mutatja a "celláknak" (cellán belül átlagolással, ha
-több opció is van). Természetesen vannak a városnak olyan részei, ahova átszállás nélkül nem lehet eljutni az
-egyetemről!
-
-### Elemzési feladat
-
-Algoritmus azon megállók megtalálására, ahova 45 perc alatt, max k átszállással el lehet jutni az I épületből este 10-es
-indulással. (Demonstráció elég k=1 vagy 2-re.)
-
-## Doc
-
-* [GTFS](doc/gtfs.md)
